# demo_07/api.yaml
openapi: 3.0.3
servers:
  - url: http://172.17.0.1:8080

paths:
  /products:
    get:
      summary: Get all products
      operationId: listProducts
      responses:
        200:
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      summary: Create a new product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        201:
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      summary: Get a product by ID
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        404:
          description: Product not found

    put:
      summary: Update a product by ID
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        200:
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      summary: Delete a product by ID
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: Product deleted successfully
        404:
          description: Product not found

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Laptop"
        price:
          type: number
          format: float
          example: 999.99
        stock:
          type: integer
          example: 50

    ProductCreate:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        price:
          type: number
          format: float
        stock:
          type: integer

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
          format: float
        stock:
          type: integer